requirements.txt
simpleitk
scikit-image
numpy
tqdm
nibabel
pandas
torchxrayvision
